Move tests from worksearch to WorkSearchScheme
