+++ openlibrary/plugins/worksearch/schemes/tests/test_works.py
@@ -1,0 +1,110 @@
+# {'Test name': ('query', fields[])}
+import pytest
+from openlibrary.plugins.worksearch.schemes.works import WorkSearchScheme
+
+
+QUERY_PARSER_TESTS = {
+    'No fields': ('query here', 'query here'),
+    'Author field': (
+        'food rules author:pollan',
+        'food rules author_name:pollan',
+    ),
+    'Field aliases': (
+        'title:food rules by:pollan',
+        'alternative_title:(food rules) author_name:pollan',
+    ),
+    'Fields are case-insensitive aliases': (
+        'food rules By:pollan',
+        'food rules author_name:pollan',
+    ),
+    'Spaces after fields': (
+        'title: "Harry Potter"',
+        'alternative_title:"Harry Potter"',
+    ),
+    'Quotes': (
+        'title:"food rules" author:pollan',
+        'alternative_title:"food rules" author_name:pollan',
+    ),
+    'Leading text': (
+        'query here title:food rules author:pollan',
+        'query here alternative_title:(food rules) author_name:pollan',
+    ),
+    'Colons in query': (
+        'flatland:a romance of many dimensions',
+        'flatland\\:a romance of many dimensions',
+    ),
+    'Spaced colons in query': (
+        'flatland : a romance of many dimensions',
+        'flatland\\: a romance of many dimensions',
+    ),
+    'Colons in field': (
+        'title:flatland:a romance of many dimensions',
+        'alternative_title:(flatland\\:a romance of many dimensions)',
+    ),
+    'Operators': (
+        'authors:Kim Harrison OR authors:Lynsay Sands',
+        'author_name:(Kim Harrison) OR author_name:(Lynsay Sands)',
+    ),
+    # LCCs
+    'LCC: quotes added if space present': (
+        'lcc:NC760 .B2813 2004',
+        'lcc:"NC-0760.00000000.B2813 2004"',
+    ),
+    'LCC: star added if no space': (
+        'lcc:NC760 .B2813',
+        'lcc:NC-0760.00000000.B2813*',
+    ),
+    'LCC: Noise left as is': (
+        'lcc:good evening',
+        'lcc:(good evening)',
+    ),
+    'LCC: range': (
+        'lcc:[NC1 TO NC1000]',
+        'lcc:[NC-0001.00000000 TO NC-1000.00000000]',
+    ),
+    'LCC: prefix': (
+        'lcc:NC76.B2813*',
+        'lcc:NC-0076.00000000.B2813*',
+    ),
+    'LCC: suffix': (
+        'lcc:*B2813',
+        'lcc:*B2813',
+    ),
+    'LCC: multi-star without prefix': (
+        'lcc:*B2813*',
+        'lcc:*B2813*',
+    ),
+    'LCC: multi-star with prefix': (
+        'lcc:NC76*B2813*',
+        'lcc:NC-0076*B2813*',
+    ),
+    'LCC: quotes preserved': (
+        'lcc:"NC760 .B2813"',
+        'lcc:"NC-0760.00000000.B2813"',
+    ),
+    # TODO Add tests for DDC
+}
+
+
+@pytest.mark.parametrize(
+    "query,parsed_query", QUERY_PARSER_TESTS.values(), ids=QUERY_PARSER_TESTS.keys()
+)
+def test_query_parser_fields(query, parsed_query):
+    s = WorkSearchScheme()
+    assert s.process_user_query(query) == parsed_query
+
+
+def test_process_user_query():
+    s = WorkSearchScheme()
+    assert s.process_user_query('test') == 'test'
+
+    q = 'title:(Holidays are Hell) authors:(Kim Harrison) OR authors:(Lynsay Sands)'
+    expect = ' '.join(
+        [
+            'alternative_title:(Holidays are Hell)',
+            'author_name:(Kim Harrison)',
+            'OR',
+            'author_name:(Lynsay Sands)',
+        ]
+    )
+    assert s.process_user_query(q) == expect
